Refactored App.test to Feeds.test
